+++ corvis/src/tracker/transformation.cpp
@@ -192,7 +192,7 @@ f_t estimate_3d_point(const aligned_vector<v2> &src, const std::vector<transform
 {
     m<Eigen::Dynamic, 4> A; A.resize(2*camera_poses.size(),4);
     m<3,4> P;
-    for (int i = 0; i< src.size(); ++i) {
+    for (size_t i = 0; i< src.size(); ++i) {
         v2 p = src[i];
         const transformation& camera = camera_poses[i];
         P.block<3,3>(0,0) = camera.Q.toRotationMatrix();
@@ -204,7 +204,7 @@ f_t estimate_3d_point(const aligned_vector<v2> &src, const std::vector<transform
 
     // calculate reprojection error at each node
     aligned_vector<f_t> errors_point;
-    for (int i = 0; i< camera_poses.size(); ++i) {
+    for (size_t i = 0; i< camera_poses.size(); ++i) {
         const transformation& G_CW = camera_poses[i];
         const feature_t& kpn = src[i];
         // predict 3d point in the camera reference
